refactor(login-page): use index file exports
