+++ apps/web/.stylelintignore
@@ -1,0 +1,1 @@
+dist
