+++ apps/web/src/components/Button/index.ts
@@ -1,0 +1,3 @@
+import Button from "./Button";
+
+export default Button;
