+++ apps/web/src/components/Card/index.ts
@@ -1,0 +1,3 @@
+import Card from "./Card";
+
+export default Card;
