+++ apps/web/src/pages/Login/LoginForm/LoginForm.tsx
@@ -1,9 +1,9 @@
 import React, { useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { z } from "zod";
-import Button from "@/components/Button/Button";
-import Card from "@/components/Card/Card";
-import Input from "@/components/Input/Input";
+import Button from "@/components/Button";
+import Card from "@/components/Card";
+import Input from "@/components/Input";
 import styles from "./LoginForm.module.css";
 
 type FlattenedErrors = z.inferFlattenedErrors<typeof FormData>;
